Add files via upload
Deleted the "magnitude" function in the "Spherical" subclass as it was not used. When calculating the magnitude of the example spherical vector in Task 2, the "norm" function in the "Vector" class was called. This worked without issue as the x, y and z values of spherical vectors were found as part of the "__init__" definition.