n-dimensional point generator/radial calculator
Defines a class that can generate an n-dimensional point with constituent components within the range of -1 to 1. Additionally, the radial of these points can be calculated using the root sum square (RSS) of the points' x, y, z, etc. coords.

As a test, points A and B (with dimensions 2 and 3, respectively) were generated, and their radial values were found. The values of the radials were consistent with the coord values.