Improvement on previous code for multiple points
Updated the point generating class to work for multiple points. The random points were generated in an n x d array where 'n' is the number of points and 'd' the dimensions. The radial calculation was also reworked to produce an n x 1 array. After creating the function 'within', the number of points found within or on the d-dimensional ball was recorded (the percentage of which was also taken).