Addition of Monte Carlo Function
A generalised Monte Carlo function was added that can find the average value, integral and variance (error) of any function. This was tested using the 2D ball example at 1000 points.

To save computational power, as the ball is of unit radius, the square root was taken out of the 'r_vector' function, and instead, the variable 'dim_sum' was returned in the code.