Gaussian Class Reworking
The previous Gaussian class only worked in 1D. The class therefore had to be rewritten to work for any specified number of dimensions. This included rewriting the functions 'integrand', 'normal' and 'normalisation'. Additionally, how the variables 'self.t' and 'self.x' were expressed was also changed.

Finally, the docstring at the start of the file was changed to include descriptions of the Monte Carlo and Gaussian classes, not just the Points class.