Editing Classes and Parallelising Code
Imported both 'time' and from mpi4py 'MPI'. Time was used to record how long the code took to run for set numbers of processors so that a percentage of parallel efficiency could be estimated.

In the 'Points' class, combined the 'r_vector' and 'ball_location' functions, and removing the while loop for more efficient computational coding.

In the 'MonteCarlo' class, combined the 'integral', 'variance' and 'calculations' functions into one, and rewriting to combine values found across the ranks using the MPI 'reduce' command. Additionally, the initialisation of the class was edited so that the number of dimensions and samples could be taken from the function the was being analysed. This helped with a pylint issue by reducing the number of arguments to an acceptable value of 5.

In the 'Gaussian' class, combined the 'normal' and 'integrand' classes into one, and removing the while loops (again for computational efficiency).